whisper-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
openai-whisper==20250625